--- python/cusignal/signaltools.py
@@ -1,48 +1,0 @@
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import cupy as cp
-
-
-def cmplx_sort(p):
-    """Sort roots based on magnitude.
-
-    Parameters
-    ----------
-    p : array_like
-        The roots to sort, as a 1-D array.
-
-    Returns
-    -------
-    p_sorted : ndarray
-        Sorted roots.
-    indx : ndarray
-        Array of indices needed to sort the input `p`.
-
-    Examples
-    --------
-    >>> from scipy import signal
-    >>> vals = [1, 4, 1+1.j, 3]
-    >>> p_sorted, indx = signal.cmplx_sort(vals)
-    >>> p_sorted
-    array([1.+0.j, 1.+1.j, 3.+0.j, 4.+0.j])
-    >>> indx
-    array([0, 2, 3, 1])
-
-    """
-    p = cp.asarray(p)
-    if cp.iscomplexobj(p):
-        indx = cp.argsort(abs(p))
-    else:
-        indx = cp.argsort(p)
-    return cp.take(p, indx, 0), indx
